requirements.txt
numpy
matplotlib
pandas
streamlit==1.34.0
streamlit-drawable-canvas
pillow